Moving create_class_instance from user to utils.
