+++ intercom/utils.py
@@ -17,7 +17,6 @@ def entity_key_from_type(type):
 
 def constantize_singular_resource_name(resource_name):
     class_name = inflection.camelize(resource_name)
-    from intercom.user import create_class_instance
     return create_class_instance(class_name)
 
 
@@ -27,3 +26,23 @@ def resource_class_to_collection_name(cls):
 
 def resource_class_to_name(cls):
     return cls.__name__.lower()
+
+
+CLASS_REGISTRY = {}
+
+def create_class_instance(class_name):
+    from intercom.traits.api_resource import Resource
+
+    if class_name in CLASS_REGISTRY:
+        return CLASS_REGISTRY[class_name]
+
+    class Meta(type):
+        def __new__(mcs, name, bases, attributes):
+            return super(Meta, mcs).__new__(mcs, str(class_name), bases, attributes)
+
+    class DynamicClass(Resource):
+        __metaclass__ = Meta
+
+    dyncls = DynamicClass()
+    CLASS_REGISTRY[class_name] = dyncls
+    return dyncls
